+++ frontend/src/pages/guest/completePage/CompletePage.tsx
@@ -1,41 +1,16 @@
-import confetti from 'canvas-confetti';
-import { useEffect, useRef } from 'react';
+import { useEffect } from 'react';
 import { useLocation, useNavigate, useParams } from 'react-router-dom';
 import Button from '../../../components/@common/buttons/button/Button';
 import { INFORMATION } from '../../../constants/messages';
 import { createGuestMainRoute, ROUTES } from '../../../constants/routes';
-import { theme } from '../../../styles/theme';
+import useConfetti from '../../../hooks/@common/useConfetti';
 import * as S from './CompletePage.styles';
 
 const CompletePage = () => {
   const { state } = useLocation();
   const navigate = useNavigate();
   const { spaceCode } = useParams();
-
-  const canvasRef = useRef<HTMLCanvasElement | null>(null);
-
-  useEffect(() => {
-    if (!canvasRef.current) return;
-
-    const myConfetti = confetti.create(canvasRef.current, {
-      resize: true,
-      useWorker: true,
-    });
-
-    const firework = () => {
-      myConfetti({
-        particleCount: 100,
-        spread: 160,
-        startVelocity: 50,
-        ticks: 200,
-        origin: { x: 0.5, y: 0.7 },
-        shapes: ['circle', 'square'],
-        gravity: 2,
-      });
-    };
-
-    firework();
-  }, []);
+  const { canvasRef, canvasStyles } = useConfetti();
 
   useEffect(() => {
     if (!state || !state.guestNickName || !state.receiver) {
@@ -44,15 +19,6 @@ const CompletePage = () => {
   }, [state, navigate]);
 
   if (!state?.receiver || !state?.guestNickName) return null;
-
-  const canvasStyles: React.CSSProperties = {
-    position: 'fixed',
-    top: '50%',
-    left: '50%',
-    transform: 'translate(-50%, -50%)',
-    width: `${theme.layout.width}px`,
-    height: `calc(100dvh - ${theme.layout.headerHeight} - ${theme.layout.padding.topBottom} - ${theme.layout.padding.leftRight})`,
-  };
 
   return (
     <>
